refactor: introduce TableDrivenRuleTest base class to consolidate table-driven test logic
